Began work on lods

+++ AbsGameProject/Maths/Heightmap.cs
@@ -55,6 +55,6 @@ public static class Heightmap
         mountains *= amplitude * 2;
         mountains = mountains * mountainStrength;
 
-        return MathF.Max(mountains, continent);
+        return (int)MathF.Max(mountains, continent);
     }
 }
